+++ README.md
@@ -99,7 +99,7 @@ bun run clean
 
 ## Secrets (local and remote)
 
-We use Bun secrets for local development and SST secrets for remote deployments. Code reads values via `getSecret(name)` (see `packages/shared-api/utils/secrets.ts`), which resolves from SST first and falls back to Bun secrets locally.
+We use Bun secrets for local development and SST secrets for remote deployments. Code reads values via `getSecret(name)` (see `packages/shared-api/utils/secret.ts`), which resolves from SST first and falls back to Bun secrets locally.
 
 Local development secrets are optional. Only configure the ones needed for the features you're working with (e.g., configure LLM provider secrets to test AI features).
