+++ apps/auth/src/better-auth.ts
@@ -6,7 +6,7 @@ import { emailOTP, organization } from "better-auth/plugins";
 import { AUTH_SECRET, AUTH_URL, LOG_LEVEL } from "@hebo/shared-api/env";
 import { COOKIE_CONFIG } from "@hebo/shared-api/lib/better-auth";
 import { createPrismaAdapter } from "@hebo/shared-api/lib/db/postgres";
-import { getSecret } from "@hebo/shared-api/utils/secrets";
+import { getSecret } from "@hebo/shared-api/utils/secret";
 import { getRootDomain } from "@hebo/shared-api/utils/url";
 
 import { PrismaClient } from "~auth/generated/prisma/client";
